Convert to .NET 5, and other refactors

+++ UndertaleModTool/ImportCodeSystem.cs
@@ -56,11 +56,14 @@ namespace UndertaleModTool
 
         public void NukeProfileGML(string codeName)
         {
+            if (!Data.ToolInfo.ProfileMode)
+                return;
+
             string path = Path.Combine(ProfilesFolder, Data.ToolInfo.CurrentMD5, "Temp", codeName + ".gml");
             if (File.Exists(path))
             {
-                //The file must be deleted to write it again properly with true decompiled output
-                //Just overwriting it isn't enough
+                // The file must be deleted to write it again properly with true decompiled output
+                // Just overwriting it isn't enough
                 File.Delete(path);
                 File.WriteAllText(path, GetDecompiledText(codeName));
             }
